% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{disc_test_data}
\alias{disc_test_data}
\title{Discretization for test dataset}
\usage{
disc_test_data(x, cutp)
}
\arguments{
\item{x}{A dataframe of test data with some numeric columns}

\item{cutp}{cut points from train dataset}
}
\value{
new test data after discretization
}
\description{
After getting cut points from train data, we could apply this cut points for discretization of test data
}
\details{
function to discrete the test data with cut points from train dataset.
}
\examples{
x=iris[c(1:40,51:90,101:140),-5]
y=iris[c(1:40,51:90,101:140),5]
testx = iris[c(41:50,91:100,141:150),-5]
v = disc_train_data(x,y)
testx_dis = disc_test_data(testx,v$cutp)
testx_dis


}
