% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{disc_train_data}
\alias{disc_train_data}
\title{Discretization for train dataset}
\usage{
disc_train_data(x,y,alpha =0.0.5)
}
\arguments{
\item{x}{A dataframe of train data with some numeric columns}

\item{y}{A dataframe or vector of categorical labels}

\item{alpha}{Significance level value, default is 0.05}
}
\value{
A list with cut points and new x dataframe
}
\description{
function of discreting the train data with supervised method and return the cut points.
}
\examples{
x=iris[c(1:40,51:90,101:140),-5]
y=iris[c(1:40,51:90,101:140),5]
testx = iris[c(41:50,91:100,141:150),-5]
v = disc_train_data(x,y)
v$discredata
v$cutp

}
