% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{myNaiveBayes}
\alias{myNaiveBayes}
\title{Naive bayes classifier with discretization and Gaussian estimation}
\usage{
myNaiveBayes(x,y,laplace = 0,discre = FALSE,alpha=0.05)
}
\arguments{
\item{x}{A dataframe of train data}

\item{y}{A dataframe or vector of categorical labels}

\item{laplace}{parametre for laplace smoothing, default is 0}

\item{discre}{paramtre to decide discretization, default is FALSE}

\item{alpha}{Significance level value for discretization, default is 0.05}
}
\value{
object for Naive bayes classifier
}
\description{
Naive bayes classifier with discretization and Gaussian estimation
}
\examples{
x=iris[c(1:40,51:90,101:140),-5]
y=iris[c(1:40,51:90,101:140),5]
testx = iris[c(41:50,91:100,141:150),-5]
m2 = myNaiveBayes(x,y)
r2 = predict_your_model(m2,testx,'class')

}
