% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{print_my_naiveBayes}
\alias{print_my_naiveBayes}
\title{Print function to see hidden information}
\usage{
print_my_naiveBayes(model)
}
\arguments{
\item{model}{object for Naive bayes classifier}
}
\value{
1 as flag
}
\description{
Print function to see hidden information
}
\examples{
x=iris[c(1:40,51:90,101:140),-5]
y=iris[c(1:40,51:90,101:140),5]
testx = iris[c(41:50,91:100,141:150),-5]
m2 = myNaiveBayes(x,y)
print_my_naiveBayes(m2)
}
