% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{predict_your_model}
\alias{predict_your_model}
\title{Naive bayes predictor}
\usage{
predict_your_model(NB_obj,new_x, pred_type = c('class','raw'),threshold = .Machine$double.eps,eps = 0)
}
\arguments{
\item{NB_obj}{object for Naive bayes classifier}

\item{new_x}{a dataframe of test dataset without labels}

\item{pred_type}{predict result, should be 'class' or 'raw'}

\item{threshold}{laplace smoothing parametre, default is .Machine$double.eps}

\item{eps}{laplace smoothing parametre, default is 0}
}
\value{
predicted result depending on pred_type
}
\description{
Naive bayes predictor
}
\examples{
x=iris[c(1:40,51:90,101:140),-5]
y=iris[c(1:40,51:90,101:140),5]
testx = iris[c(41:50,91:100,141:150),-5]
m2 = myNaiveBayes(x,y)
r2 = predict_your_model(m2,testx,'class')

}
